requirements.txt
python-gnupg==0.4.5
pyusb==1.0.2
requests
tqdm